+++ R/colorednoise.R
@@ -8,13 +8,12 @@
 #' @name colorednoise
 #' @useDynLib colorednoise, .registration = TRUE
 #' @import purrr
-#' @import dplyr
+#' @importFrom Rcpp evalCpp
+#' @rawNamespace import(data.table, except = transpose)
 #' @importFrom stats sd acf na.omit plogis
-#' @importFrom tibble tibble as_tibble
-#' @import tidyr
 NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in
 ## pipelines
 if (getRversion() >= "2.15.1") utils::globalVariables(c(".", "mean.trans",
-    "sd.trans", "noise", "timestep", "dist", "zero"))
+    "sd.trans", "noise", "timestep", "dist", "zero", "ref"))