+++ src/noise.cpp
@@ -85,11 +85,14 @@ arma::mat cor2cov(Rcpp::NumericVector sigma, Rcpp::NumericMatrix corrMatrix) {
 //' @param covMatrix A valid covariance matrix. The number of rows/columns must match the length of the mu, sigma, and phi vectors.
 //' @return A matrix with as many rows as timesteps and as many columns as mu/sigma/phi values.
 //' @examples
-//' cov <- matrix(c(0.037, 0.044, -0.048, 0.044, 0.247, -0.008, -0.047, -0.008, 0.074), nrow = 3)
+//' cov <- matrix(c(1, 0.53, 0.73, 0.53, 1, 0.44, 0.73, 0.44, 1), nrow = 3)
 //' test <- colored_multi_rnorm(100, c(0, 3, 5), c(1, 0.5, 1), c(0.5, -0.3, 0), cov)
 //' var(test)
-//' library(dplyr)
-//' test %>% as.data.frame() %>% summarize_all(.funs = c("mean", "sd", "autocorrelation"))
+//' library(data.table)
+//' as.data.table(test)[, .(V1_mean = mean(V1), V2_mean = mean(V2), V3_mean = mean(V3),
+//' V1_sd = sd(V1), V2_sd = sd(V2), V3_sd = sd(V3),
+//' V1_autocorrelation = autocorrelation(V1), V2_autocorrelation = autocorrelation(V2),
+//' V3_autocorrelation = autocorrelation(V3))]
 //' @export
 // [[Rcpp::export]]
 Rcpp::NumericMatrix colored_multi_rnorm(int timesteps, Rcpp::NumericVector mean, Rcpp::NumericVector sd, Rcpp::NumericVector phi, Rcpp::NumericMatrix covMatrix) {